Revert "solid transitions"
