--- Assets/Materials/Black.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 818a5b544ea635b45b2a6747fcfe32f3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
