+++ Assets/Scripts/ButtonTransition.cs
@@ -2,7 +2,6 @@ using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.EventSystems;
 
-
 public class ButtonTransition : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler, IPointerDownHandler, IPointerUpHandler, IPointerClickHandler
 {
     public Color32 NormalColor = Color.white;
@@ -18,37 +17,29 @@ public class ButtonTransition : MonoBehaviour, IPointerEnterHandler, IPointerExi
 
     public void OnPointerEnter(PointerEventData eventData)
     {
-       // Debug.Log("Enter");
+        Debug.Log("Enter");
         Img.color = HoverColor;
     }
 
     public void OnPointerExit(PointerEventData eventData)
     {
-     //   Debug.Log("Exit");
+        Debug.Log("Exit");
         Img.color = NormalColor;
     }
 
     public void OnPointerDown(PointerEventData eventData)
     {
-    //    Debug.Log("Down");
+        Debug.Log("Down");
         Img.color = DownColor;
-        Debug.Log(Time.time);
-        
-        FadeInOut.startFadeInTransition();
-        
-        
-       
     }
 
     public void OnPointerUp(PointerEventData eventData)
     {
-   //     Debug.Log("Up");
+        Debug.Log("Up");
     }
 
     public void OnPointerClick(PointerEventData eventData)
     {
-
-    
-        
+        Debug.Log("Click");
     }
 }
