+++ Assets/Scripts/PointerLogic.cs
@@ -12,8 +12,6 @@ public class PointerLogic : MonoBehaviour
 
     private LineRenderer LineRender = null;
 
-
-
     private void Awake()
     {
         LineRender = GetComponent<LineRenderer>();
@@ -24,9 +22,6 @@ public class PointerLogic : MonoBehaviour
         UpdateLine();
     }
 
-
-    
-    
     private void UpdateLine()
     {
         PointerEventData data = InputModule.GetData();
