+++ Assets/SteamVR_Resources/Resources/SteamVR_Settings.asset
@@ -24,8 +24,8 @@ MonoBehaviour:
   autoEnableVR: 1
   legacyMixedRealityCamera: 1
   mixedRealityCameraPose:
-    actionPath: /actions/mixedreality/in/ExternalCamera
-    needsReinit: 0
+    actionPath: /actions//in/ExternalCamera
+    needsReinit: 1
   mixedRealityCameraInputSource: 11
   mixedRealityActionSetAutoEnable: 1
   previewHandLeft: {fileID: 0}
