+++ ProjectSettings/EditorBuildSettings.asset
@@ -4,16 +4,7 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes:
-  - enabled: 1
-    path: Assets/Scenes/Env1.unity
-    guid: 0a70ec000ee0e9d4681cc165dd5c7e4e
-  - enabled: 1
-    path: Assets/Scenes/Env2.unity
-    guid: 15426fa7abcf79c46b89f56c678980ed
-  - enabled: 1
-    path: Assets/Scenes/Env3.unity
-    guid: 95f7f1622d15b344e9f49248db3e8212
+  m_Scenes: []
   m_configObjects:
     Unity.XR.OpenVR.Settings: {fileID: 11400000, guid: d001f7466033a634c8f4aa37639031f3, type: 2}
     com.unity.xr.management.loader_settings: {fileID: 11400000, guid: 23fecdd259187204db46bc1f95547183, type: 2}
