+++ unityProject.vrmanifest
@@ -5,7 +5,7 @@
       "app_key": "application.generated.unity.testvr.exe",
       "launch_type": "url",
       "url": "steam://launch/",
-      "action_manifest_path": "C:\\Users\\jiang\\Desktop\\ICS\\Languages\\C Language\\C#\\Unity Projects\\CalmVR\\Assets\\StreamingAssets\\SteamVR\\actions.json",
+      "action_manifest_path": "C:\\Users\\User\\Documents\\GitHub\\CalmVR\\Assets\\StreamingAssets\\SteamVR\\actions.json",
       "strings": {
         "en_us": {
           "name": "Test VR [Testing]"
